+++ quantfit/__init__.py
@@ -3,5 +3,5 @@
 from quantfit.gpufit import FitReport, check_fit
 from quantfit.spec import DEFAULT_SPEC, QuantSpec
 
-__version__ = "0.3.0"
+__version__ = "0.4.0"
 __all__ = ["FitReport", "check_fit", "QuantSpec", "DEFAULT_SPEC", "__version__"]